- Implementors of the interface "ValueSupplier" has to return a double value upon request, signaling an end point with .hasNext().
- Implementors of "ValueSupplierHost" are currently Sequence and
PatternCombiner, where the former is a list-based pattern, and the
latter combines patterns to one single event. A ValueSupplierHost is
typically a container for ValueSuppliers, and has methods for adding,
replacing, and removing ValueSuppliers.

- The ValueSupplierField is a UI element which lets the user choose
between different kinds of patterns. It is currently a part of
SequenceUI's and PatternCombinerUI's cells. A ValueSupplierField has to
register with a ValueSupplierHost to work properly. A ValueSupplierField
doesn't store state (as should be the case with UI), but holds a
reference to a ValueSupplier which is already a part of a larger
structure like PatternCombiner or Sequence.

- The default font for the project is moved to TestUI, the temporary
starting point for the program.

- Some older code is removed from the project. More code will be removed
later.

- The program is in a working state as of now, but is limited to sending
midi to the default port.

- A sequence can now be selected as a element in a sequence, ad
infinitum.